uploads: Cache boto client in the module, not class, and be writable.
The `get_boto_client` method set `signature_version=botocore.UNSIGNED`
which meant that it could not be used to make authenticated requests,
only unauthenticated requests (e.g. fetching files in the avatars
bucket, no-op signing public access URLs).  It could not be used for
the module-level `get_signed_upload_url` because it was in the class,
as well as being write-only.

The `get_signed_upload_url` code is called for every S3 file upload
request, and is thus in the hot path.  The boto3 client caching
optimization is thus potentially useful as a performance optimization.

Move `get_boto_client` to the module, and remove the
`signature_version=botocore.UNSIGNED` configuration flag which makes
it unauthenticated.

The `generate_presigned_url` call in
`construct_public_upload_url_base` now returns a URL with
`AWSAccessKeyId=...&Signature=...&Expires=` parameters, but those are
stripped out by the urllib parsing.  However, the call in
`get_export_tarball_url` now also requires stripping query parameters
to expose the public URL.