models: Add tus_file_id field to the Attachment model.
We are going to use tusd to support the chunk upload. As of now, the
latest stable version of tusd v1.11.0 doesn't support sending custom
responses but it do send tus_file_id, so we will use that tus_file_id
to get the upload path of the file, implementing a new endpoint.