upload: Use @uppy/tus at client side to upload a file.
This commit replaces the previoud XHRUpload plugin with Tus
to utilize new chunk upload feature. Currently chunk size is 25MB,
but later we can take this from server.

Tusd server at backend doesn't support sending custom path in the
response so we are utilizing an endpoint which accepts tus_file_id
in url param and returns the path of file. Later we can send the path in
response itself and remove the extra endpoint.