subscription: Include archived streams in unsubscribed streams list.
`is_archived` field is added to the stream and types.

Include a new `archived_streams` client capability, to allow clients
to access data on archived streams, without breaking
backwards-compatibility for existing clients that don't know how to
handle these.

Also, included `exclude__archived` parameter to `/get-streams`,
which defaults to `true` as basic clients may not be interested
in archived streams.