stream_settings: Add option to send notification to newly added users.
This commit adds a checkbox in stream settings and stream
creation form to choose whether to send a notification
to users when added to a channel.

If more than 100 users are being added at once, option
becomes unchecked and disabled.
If the number of users being added goes below 100 after being higher,
the option becomes re-enabled, but remains unchecked.

Fixes #31189