message_send: Fix tests to use json encoded 'to'.
OpenApi docs requires the `to` param to be Json encoded,
but the tests were using raw strings. This commit fixes the tests.
Despite OpenApi docs requiring the `to` param to be Json encoded,
the endpoint still accepts the `to` param as a raw string.
This is undesireable, but fixing it would render older bots that
take advantage of this feature unusable.