billing: Highlight manual billing errors for admins.
Limit decreasing the number of licenses for the next billing period
to be less than the currently used licenses in the realm. If the
customer is exempt from license number checks, then this limit is
not applied.

Also, highlight manual billing errors so that the billing admin is
aware of the issue.

Current licenses can be under the seat count when a guest user is
changed to a non-guest user.

Next billing period licenses can be under the seat cout when a user
joins with a currently available, purchased license after a billing
admin has decreased the number of licenses set for the next billing
period.