popovers: Add stream-info-popover when channel pill is clicked.
Shows a stream info popover whenever a channel pill is clicked (eg. in the UI for adding subscribers to a channel).The stream info popover contains channel icon and name, channel description and a link to channel settings. Also, it renames "popover-menu-inner-list-item" to "popover-menu-list-item" in "left_sidebar_stream_actions_popover.hbs" as instructed.
Fixes #30567.